+++ NEWS.md
@@ -1,5 +1,7 @@
 # dials (development version)
 
+* Renamed `margin()` to `svm_margin()` (#85). We've updated the tests for `margin()` and `svm_margin()` and added the `test_deprecated.R` test script. (@gabrielodom and @gralgomez, #85)
+
 # dials 0.0.4
 
 ## New parameter functions: